backend/requirements.txt
fastapi
uvicorn
pillow
fitz
pymupdf
easyocr
torch
ultralytics
opencv-python
numpy
python-multipart
pydantic